+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ai-sdk/rsc
 
+## 2.0.0-beta.104
+
+### Patch Changes
+
+- ai@6.0.0-beta.103
+
+## 2.0.0-beta.103
+
+### Patch Changes
+
+- ai@6.0.0-beta.102
+
+## 2.0.0-beta.102
+
+### Patch Changes
+
+- ai@6.0.0-beta.101
+
+## 2.0.0-beta.101
+
+### Patch Changes
+
+- Updated dependencies [8370068]
+  - ai@6.0.0-beta.100
+
 ## 2.0.0-beta.100
 
 ### Patch Changes